Unit Tests added and mock data is now in a seperate folder
